refactor: organized base components folder
