+++ src/components/cart/base/DiscountCode.tsx
@@ -1,5 +1,5 @@
 import Button from '@/components/base/Button'
-import Input from '@/components/base/Input'
+import Input from '@/components/base/input/Input'
 import { useCartStore } from '@/stores/cartStore'
 import { DiscountCodeSchema, TDiscountCodeSchema } from '@/lib/validators'
 import { validateDiscountCode } from '@/services/coupons'
